Added new content from RN 2021.1 Confluence document - open questions for Shay
Added sections include:

Operations and Configuration Changes (and all sub-sections)
Open question for Shay about a link to a Confluence page. Does the Confluence page need to be converted to our website format before being included in the RN 2021.1? In other words, it doesn't seem like we can link from a documentation page to a Confluence page.
This is the Confluence URL: https://sqream.atlassian.net/wiki/spaces/DOC/pages/822902789/How+to+connect+sqream+server+to+Cloudera+Hadoop+with+kerberos