Created get_license_info
Content based on 2021.1 RN Confluence doc. This update is the attempt made from GitHub Desktop explorer. For some reason, adding the file in GitHub repo did not display it in Explorer, even after fetching.

This could be because I could not change the name of the branch from v2021.2 to v2021.1. Confirmed that the RN file in branch v2021.2 are the 2021.1 RN files.