Update installing_sqream_with_binary.rst
Added in Overview "this needs to be done on all servers."
Added after 'For each SQream service configuration file, do the following:" -  NOTE: if you are running sqeam on more than one server, configure serverpicker and metaatserver services to start only one of them. If the first server is running metadaterver, then in /etc/sqream/sqream1_config.json of the second server, metadataServerIp value shuld point to the IP of the server that is running metadatserver service.