Deleted Setup folder (sqream_docs\guides\operations\setup).
Folder was not necessary because it contains files that are either outdated or are irrelevant:

before_you_begin - irrelevant. The content on this page was moved to Getting Started page.
index - irrelevant
installing_monit - this page already exists in Operations.
local_docker - there is an updated Docker page in the Operations menu. I've backed up this file.
recommended_configuration - there is an updated page in the Operations menu.