Replaced VARCHAR with TEXT
Replaced *most* instances of VARCHAR with TEXT. I didn't replace all of them because I had some open questions for Shachar (about 20). I am going to merge this now into branch 2020-1 to test if it worked, and will merge it again after updating it in light of Shachar's comments. I will then merge it on the remaining branches.